requirements.txt
# Core ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Natural Language Processing
nltk>=3.8.0

# Model Persistence
joblib>=1.3.0

# API Integration
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Progress Display
tqdm>=4.66.0

# Text Processing
regex>=2023.0.0

# Optional: For advanced NLP (uncomment if needed)
# transformers>=4.30.0
# torch>=2.0.0
